Remove unsupported stuff from {pre,post}install scripts

+++ dist-assets/linux/before-remove.sh
@@ -2,10 +2,6 @@
 set -eu
 
 echo "Running before-remove.sh"
-# SIGTERM for some reason causes the app to crash sometimes and SIGINT works as expected.
-pkill -2 -x "mullvad-gui" || true
-sleep 0.5
-pkill -9 -x "mullvad-gui" || true
 
 is_number_re='^[0-9]+$'
 # Check if we're running during an upgrade step on Fedora
@@ -19,10 +15,12 @@ if [[ "$1" == "upgrade" ]]; then
 fi
 
 # the user might've disabled or stopped the service themselves already
-systemctl stop mullvad-daemon.service || true
-systemctl disable mullvad-daemon.service || true
-systemctl stop mullvad-early-boot-blocking.service || true
-systemctl disable mullvad-early-boot-blocking.service || true
+if which systemctl &> /dev/null; then
+    systemctl stop mullvad-daemon.service || true
+    systemctl disable mullvad-daemon.service || true
+    systemctl stop mullvad-early-boot-blocking.service || true
+    systemctl disable mullvad-early-boot-blocking.service || true
+fi
 
 /opt/Mullvad\ VPN/resources/mullvad-setup reset-firewall || echo "Failed to reset firewall"
 /opt/Mullvad\ VPN/resources/mullvad-setup remove-device || echo "Failed to remove device from account"
